+++ src/pybamm/models/full_battery_models/base_battery_model.py
@@ -1228,9 +1228,12 @@ class BaseBatteryModel(pybamm.BaseModel):
         elif options["intercalation kinetics"] == "MSMR":
             return pybamm.kinetics.MSMRButlerVolmer
 
-    def get_inverse_intercalation_kinetics(self):
-        if self.options["intercalation kinetics"] == "symmetric Butler-Volmer":
+    def get_inverse_intercalation_kinetics(self, domain):
+        options = getattr(self.options, domain)
+        if options["intercalation kinetics"] == "symmetric Butler-Volmer":
             return pybamm.kinetics.InverseButlerVolmer
+        elif options["intercalation kinetics"] == "linear":
+            return pybamm.kinetics.InverseLinear
         else:
             raise pybamm.OptionError(
                 "Inverse kinetics are only implemented for symmetric Butler-Volmer. "